+++ FASTDOOM/i_ibm.h
@@ -4,7 +4,7 @@
 #include "options.h"
 
 extern unsigned int ticcount;
-extern unsigned long long mscount;
+extern unsigned int mscount;
 extern unsigned int fps;
 
 extern unsigned short *currentscreen;
